requirements.txt
bidict==0.23.1
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.10.1
geventhttpclient==2.3.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.31.8
MarkupSafe==3.0.1
msgpack==1.1.0
psutil==6.0.0
PyJWT==2.9.0
python-engineio==4.9.1
python-socketio==5.11.4
pyzmq==26.2.0
requests==2.32.3
setuptools==75.1.0
simple-websocket==1.1.0
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
zope.event==5.0
zope.interface==7.1.0
